Cleanup of updater and arguments, commented out obsolete code, docstring editing
